refactor(advanced): debounced Search 훅스 분리
