+++ src/advanced/features/search/index.ts
@@ -1,0 +1,1 @@
+export { useDebouncedSearch } from './lib/useDebouncedSearch';
